experiments/pdf_extraction/requirements.txt
pandas>=2.2.0
matplotlib>=3.8.2
pdfplumber>=0.10.3
PyMuPDF>=1.23.8
tika>=2.6.0
pytesseract>=0.3.10
Pillow>=10.2.0
psutil>=5.9.8
numpy>=1.26.4
reportlab>=4.1.0
lorem>=0.1.1 